more int->size_t warnings

+++ lib/dpdk/iceBoardStandard.hpp
@@ -224,7 +224,7 @@ inline bool iceBoardStandard::handle_lost_samples(int64_t lost_samples) {
 
     // TODO this could be made more efficient by breaking it down into blocks of memsets.
     while (lost_samples > 0) {
-        if (unlikely(lost_sample_location * sample_size == out_buf->frame_size)) {
+        if (unlikely((size_t)(lost_sample_location * sample_size) == out_buf->frame_size)) {
             if (!advance_frame(temp_seq)) {
                 return false;
             }
@@ -246,10 +246,10 @@ inline bool iceBoardStandard::copy_packet(struct rte_mbuf* mbuf) {
     // Note this assumes that frame_size is divisable by samples_per_packet,
     // or the assert below will fail.
     int64_t sample_location = cur_seq - get_fpga_seq_num(out_buf, out_frame_id);
-    assert(sample_location * sample_size <= out_buf->frame_size);
+    assert((size_t)(sample_location * sample_size) <= out_buf->frame_size);
 
     // Check if we are at the end of the current frame
-    if (unlikely(sample_location * sample_size == out_buf->frame_size)) {
+    if (unlikely((size_t)(sample_location * sample_size) == out_buf->frame_size)) {
         // If there are no new frames to fill, we are just dropping the packet
         if (!advance_frame(cur_seq))
             return false;
